Updated logic for repeating values
I realised in my sleep that i was repeating the low values 8 times, mid values 4 times and the high values only once and the clockdivider for the high values only multiplies by 2 instead of for, for the high values. so i woke up and got right at it. My dentist sais i grind in my sleep, she Real AF for that.